# ⚡️ Speed up type checking