--- a/src/assets/courses/full-courses.json.d.ts
+++ b/src/assets/courses/full-courses.json.d.ts
@@ -1,0 +1,3 @@
+declare const json: Readonly<Record<string, readonly CornellCourseRosterCourse[]>>;
+
+export default json;