--- a/src/requirements/decorated-requirements.json.d.ts
+++ b/src/requirements/decorated-requirements.json.d.ts
@@ -1,0 +1,4 @@
+import { DecoratedRequirementsJson } from './types';
+
+const decoratedRequirementJson: DecoratedRequirementsJson;
+export default decoratedRequirementJson;